extract updateDisplay and Time
